refactor(git): extract planning workspace input validation
